Generate CA codes as int8 types, no need for them to be longs!

+++ gnss_sdr/include/generateCAcode.py
@@ -81,7 +81,7 @@ def generateCAcode(PRN):
 
   return CAcode
 
-caCodes = np.empty((32, 1023))
+caCodes = np.empty((32, 1023), dtype=np.int8)
 for PRN in range(32):
   caCodes[PRN][:] = generateCAcode(PRN)
